requirements.txt
bandit==1.7.9
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
coverage==7.6.1
flake8==7.1.1
idna==3.10
iniconfig==2.1.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
myql==1.2.7
oauthlib==3.3.1
packaging==25.0
pbr==6.1.1
pluggy==1.6.0
pyaml==24.12.1
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.19.2
pytest==8.3.3
python-dotenv==1.0.1
PyYAML==6.0.2
rauth==0.7.3
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
six==1.17.0
stevedore==5.4.1
stringcase==1.2.0
urllib3==2.5.0
yahoo-oauth==2.1.0
yfpy==16.0.3
-e .